vrrp: extract advert/event send to instance_send.go
Pure code-motion split of pkg/vrrp/instance.go (#5661). Move the
outbound advertisement/event send path into a cohesive sibling file,
package vrrp:

  emitEvent, sendAdvert, sendPacket, sendPacketIPv6.

Declarations moved verbatim — no rename of exported symbols, no
logic/timing/const change, and no reorder of any goroutine, defer,
channel, or mutex operation (whole functions moved intact). Imports
pruned by goimports; gofmt clean. Builds green. instance.go now under
the 2000-LOC threshold.
