+++ pkg/vrrp/instance.go
@@ -1501,227 +1501,6 @@ func (vi *vrrpInstance) scheduleVIPRemoveReconcile() {
 	}()
 }
 
-// emitEvent sends a state change event to the manager's event channel.
-func (vi *vrrpInstance) emitEvent() {
-	evt := VRRPEvent{
-		Interface: vi.cfg.Interface,
-		Family:    vi.cfg.Family,
-		GroupID:   vi.cfg.GroupID,
-		State:     vi.getState(),
-		VIPs:      vi.cfg.VirtualAddresses,
-	}
-	select {
-	case vi.eventCh <- evt:
-	default:
-		// Drop if channel full — warn unless we're shutting down.
-		select {
-		case <-vi.stopCh:
-		default:
-			slog.Warn("vrrp: event channel full, dropping event",
-				"key", vi.key(), "state", evt.State)
-			if vi.onEventDrop != nil {
-				vi.onEventDrop()
-			}
-		}
-	}
-}
-
-// sendAdvert sends a VRRPv3 advertisement with the given priority.
-func (vi *vrrpInstance) sendAdvert(priority int) {
-	hasIPv6 := false
-	var v4Addrs, v6Addrs []net.IP
-	for _, vip := range vi.cfg.VirtualAddresses {
-		addr := vip
-		if idx := strings.Index(addr, "/"); idx >= 0 {
-			addr = addr[:idx]
-		}
-		ip := net.ParseIP(addr)
-		if ip == nil {
-			continue
-		}
-		if ip.To4() != nil {
-			v4Addrs = append(v4Addrs, ip.To4())
-		} else {
-			v6Addrs = append(v6Addrs, ip.To16())
-			hasIPv6 = true
-		}
-	}
-
-	// Send IPv4 advertisement if we have any IPv4 VIPs.
-	if len(v4Addrs) > 0 {
-		maxAdvert := uint16(vi.cfg.AdvertiseInterval / 10) // milliseconds → centiseconds
-		pkt := &VRRPPacket{
-			VRID:         uint8(vi.cfg.GroupID),
-			Priority:     uint8(priority),
-			MaxAdvertInt: maxAdvert,
-			IPAddresses:  v4Addrs,
-		}
-		if err := vi.sendPacket(pkt, false); err != nil {
-			slog.Debug("vrrp: failed to send IPv4 advert",
-				"key", vi.key(), "err", err)
-		}
-	}
-
-	// Send IPv6 advertisement if we have any IPv6 VIPs.
-	if hasIPv6 && len(v6Addrs) > 0 {
-		maxAdvert := uint16(vi.cfg.AdvertiseInterval / 10) // ms → centiseconds
-		pkt := &VRRPPacket{
-			VRID:         uint8(vi.cfg.GroupID),
-			Priority:     uint8(priority),
-			MaxAdvertInt: maxAdvert,
-			IPAddresses:  v6Addrs,
-		}
-		if err := vi.sendPacket(pkt, true); err != nil {
-			slog.Debug("vrrp: failed to send IPv6 advert",
-				"key", vi.key(), "err", err)
-		}
-	}
-}
-
-// sendPacket sends a VRRP advertisement via the per-instance raw socket.
-func (vi *vrrpInstance) sendPacket(pkt *VRRPPacket, isIPv6 bool) error {
-	if isIPv6 {
-		return vi.sendPacketIPv6(pkt)
-	}
-	if vi.rawConn == nil {
-		return nil
-	}
-
-	srcIP := vi.getLocalIP()
-	if srcIP == nil {
-		// Lazy resolve: the interface may not have had an address at socket
-		// open time, or the addr-watcher invalidated a flushed source (#2528).
-		// Skip VIPs — must send from the primary/base address. The atomic
-		// setLocalIP makes this run-loop write race-clean against the
-		// receiver-goroutine reads (#2258).
-		srcIP = vi.resolveLocalIPv4(vi.vipAddrSet())
-		if srcIP != nil {
-			vi.setLocalIP(srcIP)
-		}
-		if srcIP == nil {
-			return fmt.Errorf("no IPv4 address on %s", vi.cfg.Interface)
-		}
-	}
-
-	dstIP := net.IPv4(224, 0, 0, 18)
-
-	data, err := pkt.Marshal(false, srcIP, dstIP)
-	if err != nil {
-		return err
-	}
-
-	hdr := &ipv4.Header{
-		Version:  4,
-		Len:      20,
-		TotalLen: 20 + len(data),
-		TTL:      255,
-		Protocol: vrrpProto,
-		Src:      srcIP,
-		Dst:      dstIP,
-	}
-
-	if err := vi.rawConn.SetMulticastInterface(vi.iface); err != nil {
-		return fmt.Errorf("set multicast interface: %w", err)
-	}
-
-	cm := &ipv4.ControlMessage{
-		IfIndex: vi.iface.Index,
-	}
-
-	if err := vi.rawConn.WriteTo(hdr, data, cm); err != nil {
-		return fmt.Errorf("writeto: %w", err)
-	}
-
-	return nil
-}
-
-// sendPacketIPv6 sends a VRRPv3 IPv6 advertisement.
-// Source: link-local address, Destination: ff02::12, Hop Limit: 255.
-func (vi *vrrpInstance) sendPacketIPv6(pkt *VRRPPacket) error {
-	if vi.ipv6Conn == nil {
-		return nil
-	}
-
-	srcIP := vi.getLocalIPv6()
-	if srcIP == nil {
-		// Lazy resolve: deterministically select the lowest link-local
-		// address. This happens when the interface didn't have a
-		// link-local at openSocket() time (e.g. DAD still running). The
-		// atomic setLocalIPv6 makes this run-loop write race-clean against
-		// the receiver-goroutine reads (#2258).
-		resolved := vi.resolveIPv6LinkLocal(vi.vipAddrSet())
-		if resolved != nil {
-			vi.setLocalIPv6(resolved)
-			srcIP = resolved
-			slog.Info("vrrp: late-resolved IPv6 link-local address",
-				"key", vi.key(), "addr", srcIP)
-		}
-		if srcIP == nil {
-			slog.Warn("vrrp: no link-local IPv6 address, skipping IPv6 advert",
-				"key", vi.key(), "interface", vi.cfg.Interface)
-			return fmt.Errorf("no link-local IPv6 address on %s", vi.cfg.Interface)
-		}
-	}
-
-	dstIP := net.ParseIP("ff02::12")
-
-	// RFC 5798 §5.2.9 / §6.1: an IPv6 VRRP advertisement lists ALL of the
-	// virtual router's addresses, and the FIRST IPvX address in the payload
-	// MUST be the virtual router's link-local. srcIP is exactly that
-	// link-local (it is also the outer source and the pseudo-header checksum
-	// source below, #2644), so prepend it ahead of the configured global
-	// VIPs the caller placed in pkt.IPAddresses. Marshal derives the
-	// "Count IPvX Addr" wire field from len(IPAddresses) (packet.go), so
-	// prepending here bumps the count in lockstep — no separate count math.
-	// Build a fresh slice rather than mutating the caller's v6Addrs so a
-	// retried send does not prepend twice.
-	llFirst := make([]net.IP, 0, len(pkt.IPAddresses)+1)
-	llFirst = append(llFirst, srcIP)
-	llFirst = append(llFirst, pkt.IPAddresses...)
-	pkt.IPAddresses = llFirst
-
-	// The pseudo-header checksum inside data is computed over srcIP. The
-	// outer IPv6 source MUST equal srcIP or the receiver's checksum (which
-	// is validated against the actual outer source) mismatches and the
-	// advert is dropped (#2644).
-	data, err := pkt.Marshal(true, srcIP, dstIP)
-	if err != nil {
-		return err
-	}
-
-	// Don't set Zone — the socket already has IPV6_MULTICAST_IF bound
-	// to the interface. Setting Zone on the destination can cause EINVAL
-	// on some kernels when the socket option is already set.
-	dst := &net.IPAddr{
-		IP: dstIP,
-	}
-
-	// Pin the outer source via IPV6_PKTINFO to the SAME address the
-	// checksum was computed over (#2644). Without this the kernel selects
-	// the outer source independently (RFC 6724) on the wildcard-bound
-	// socket and can pick a different link-local than srcIP. IfIndex
-	// scopes the link-local source to the VRRP interface (link-locals are
-	// zone-scoped) and keeps the egress interface deterministic.
-	cm := &ipv6.ControlMessage{
-		Src:     srcIP,
-		IfIndex: vi.iface.Index,
-	}
-	if vi.ipv6Send == nil {
-		// Defensive: a wired ipv6Conn without a send seam should not
-		// happen (openSocket sets both together), but fall back to the
-		// raw conn rather than panic.
-		if _, err := vi.ipv6Conn.WriteTo(data, dst); err != nil {
-			return fmt.Errorf("ipv6 writeto: %w", err)
-		}
-		return nil
-	}
-	if err := vi.ipv6Send(data, cm, dst); err != nil {
-		return fmt.Errorf("ipv6 writeto: %w", err)
-	}
-
-	return nil
-}
-
 // addVIPs adds virtual IP addresses to the interface via netlink.
 // vipActuationResult reports the outcome of an attempt to add the instance's
 // virtual IP set on the interface (#5082). It lets becomeMaster/ReconcileVIPs
